+++ torch_points3d/datasets/base_dataset.py
@@ -210,9 +210,24 @@ class BaseDataset:
         if precompute_multi_scale:
             self.set_strategies(model)
 
+        self._set_has_labels()
+
+    @property
+    def has_train_loader(self):
+        return hasattr(self, "_train_loader")
+
     @property
     def has_val_loader(self):
         return hasattr(self, "_val_loader")
+
+    def _set_has_labels(self):
+        for loader in self._loaders:
+            sample = loader.dataset[0]
+            has_labels = False
+            if hasattr(sample, "y"):
+                has_labels = sample.y is not None
+            setattr(loader, "has_labels", has_labels)
+            setattr(loader.dataset, "has_labels", has_labels)
 
     @property
     def has_test_loaders(self):
@@ -276,11 +291,22 @@ class BaseDataset:
         return self._test_loaders
 
     @property
+    def _loaders(self):
+        loaders = []
+        if self.has_train_loader:
+            loaders += [self.train_dataloader]
+        if self.has_val_loader:
+            loaders += [self.val_dataloader]
+        if self.has_test_loaders:
+            loaders += self.test_dataloaders
+        return loaders
+
+    @property
     def num_test_datasets(self):
         return len(self._test_dataset) if self._test_dataset else 0
 
     @property
-    def test_datatset_names(self):
+    def _test_datatset_names(self):
         if self.test_dataset:
             return [d.name for d in self.test_dataset]
         else:
@@ -288,10 +314,34 @@ class BaseDataset:
 
     @property
     def available_stage_names(self):
-        out = self.test_datatset_names
+        out = self._test_datatset_names
         if self.has_val_loader:
             out += [self._val_dataset.name]
         return out
+
+    @property
+    def available_dataset_names(self):
+        return ["train"] + self.available_stage_names
+
+    def get_raw_data(self, stage, idx, **kwargs):
+        assert stage in self.available_dataset_names
+        dataset = self.get_dataset(stage)
+        if hasattr(dataset, "get_raw_data"):
+            return dataset.get_raw_data(idx, **kwargs)
+        else:
+            raise Exception("Dataset {} doesn t have a get_raw_data function implemented".format(dataset))
+
+    def has_labels(self, stage: str) -> bool:
+        """ Tests if a given dataset has labels or not
+
+        Parameters
+        ----------
+        stage : str
+            name of the dataset to test
+        """
+        assert stage in self.available_dataset_names
+        dataset = self.get_dataset(stage)
+        return dataset.has_labels
 
     @property
     def has_fixed_points_transform(self):